feat: Add LIMIT_GLIBC_MALLOC_ARENAS to package (#2)
This feature is already in the development source code, however it is
currently needed for some deployments and is thus being fast-tracked
into the package as a revision. This patch will be removed after the
next source code release (which will include the feature)

Also added is a critical fix for the read cache and some windows client
related additions.