fix: Fix nfs-ganesha package FSAL library being overwritten (#3)
In 4.6.0-1, the installation of the FSAL was changed not to be hardcoded
to /usr/lib/ganesha/, but instead to be /usr/lib/*/ganesha, where the *
could be /usr/lib/x86_64-linux-gnu/ or simply /usr/lib/ganesha/, for
example. However, a symlink in the .links file was linking
/usr/lib/x86_64-linux-gnu/libfsalsaunafs.so incorrectly to
/usr/lib/ganesha/libfsalsaunafs.so, causing it to be overwritten if it
the libfsalsaunafs.so was installed in x86_64-linux-gnu.

This commit makes the FSAL packaging more aligned with Ubuntu and
Debain, which is to copy all libfsalsaunafs.so files to the same
wildcard path in .install, and removing the .links file.

An example of FSAL's building on Ubuntu can be found here:
https://git.launchpad.net/ubuntu/+source/nfs-ganesha/tree/debian/nfs-ganesha-gpfs.install?h=applied/ubuntu/jammy