fix: Fix SaunaFS uraft `ExecStopPost` directive
Previously, the `ExecStopPost` directive in the uraft systemd unit file
invoked `saunafs-uraft-helper demote`. This command not only removed
the floating IP but also restarted the master server in shadow mode,
as typically done during a leader demotion.

However, this behavior is inappropriate when stopping uraft service:
- It incorrectly restarts the master process, even on elector nodes
  where the master should never run.
- It introduces unnecessary delays, as stopping uraft leads to a
  restart of another service.

This commit changes the `ExecStopPost` action to call `drop-ip`,
which safely removes the floating IP without triggering unneeded
restarts. This is the correct behavior for metadata nodes when uraft
is stopped.

Signed-off-by: Crash <crash@leil.io>