fix(uRaft): Prevent electors from starting MDS (#10)
Previously, the `saunafs-uraft` service started `saunafs-ha-master`
unconditionally, without considering the `URAFT_ELECTOR_MODE` setting.
As a result, when uRaft started, the master service was also started
on elector nodes, which should only participate in leader elections
and not start the master process.

This behavior required manual intervention to stop the master process
on electors or custom changes to the `saunafs-uraft` systemd unit to
decouple it from `saunafs-ha-master`.

This commit modifies saunafs-ha-master service systemd to use
`saunafs-uraft-helper` script to manage the master service based on the
`URAFT_ELECTOR_MODE` setting in the uRaft configuration file.

Signed-off-by: Crash <crash@leil.io>